doc: add items hooks
